Move Bullet, Rocket to projectiles.py
