+++ ship.py
@@ -2,7 +2,8 @@ import pygame
 from pygame.math import Vector2
 from pygame import Color
 import math
-from physics import Disk, Bullet, Rocket
+from physics import Disk
+from projectiles import Bullet, Rocket
 from camera import Camera
 from enum import Enum
 import random
